+++ samples/Gallery/Pages/TransitioningContentControlPage.fs
@@ -1,96 +1,26 @@
 namespace Gallery
 
 open System
-open System.Collections.Generic
 open System.Diagnostics
-open System.Threading
-open System.Threading.Tasks
-open Avalonia
 open Avalonia.Animation
 open Avalonia.Controls
 open Avalonia.Layout
 open Avalonia.Media
-open Avalonia.Media.Imaging
-open Avalonia.Styling
-open Avalonia.VisualTree
 open Fabulous.Avalonia
 open Fabulous
+open Avalonia.Animation.Easings
 
-type PageTransition(displayTitle: string) =
-    let _transition: IPageTransition = null
-
-    member val Transition = _transition with get, set
-
-    member val DisplayTitle = displayTitle
-
-    override this.ToString() = this.DisplayTitle
-
-type CustomTransition(duration) =
-
-    let mutable _duration: TimeSpan = Unchecked.defaultof<_>
-    member val Duration = _duration with get, set
-
-    member this.GetVisualParent(from: Visual, to': Visual) =
-        let p1 = from.GetVisualParent()
-        let p2 = to'.GetVisualParent()
-
-        if p1 <> null && p2 <> null && p1 <> p2 then
-            raise(ArgumentException("Controls for PageSlide must have same parent."))
-
-        if p1 = null then
-            raise(InvalidOperationException("Cannot determine visual parent."))
-        else
-            p1
-
-    interface IPageTransition with
-        member this.Start(from: Visual, to': Visual, forward: bool, cancellationToken: CancellationToken) =
-            task {
-                if cancellationToken.IsCancellationRequested then
-                    ()
-                else
-                    let tasks = List<Task>()
-                    let parent = this.GetVisualParent(from, to')
-                    let scaleProperty = ScaleTransform.ScaleYProperty
-
-                    if from <> null then
-                        let animation = Animation()
-                        let keyFrame = KeyFrame()
-                        keyFrame.Setters.AddRange [ Setter(Property = scaleProperty, Value = 1.) ]
-                        keyFrame.Cue <- Cue(0.)
-                        let keyFrame' = KeyFrame()
-                        keyFrame'.Setters.AddRange [ Setter(Property = scaleProperty, Value = 0.) ]
-                        keyFrame'.Cue <- Cue(1.)
-                        animation.Children.AddRange([ keyFrame; keyFrame' ])
-                        animation.Duration <- this.Duration
-                        tasks.Add(animation.RunAsync(from, cancellationToken))
-
-                    if to' <> null then
-                        let animation = Animation()
-                        let keyFrame = KeyFrame()
-                        keyFrame.Setters.AddRange [ Setter(Property = scaleProperty, Value = 0.) ]
-                        keyFrame.Cue <- Cue(0.)
-                        let keyFrame' = KeyFrame()
-                        keyFrame'.Setters.AddRange [ Setter(Property = scaleProperty, Value = 1.) ]
-                        keyFrame'.Cue <- Cue(1.)
-                        animation.Children.AddRange([ keyFrame; keyFrame' ])
-                        animation.Duration <- this.Duration
-                        tasks.Add(animation.RunAsync(to', cancellationToken))
-
-                    do! Task.WhenAll(tasks)
-
-                    if from <> null && not cancellationToken.IsCancellationRequested then
-                        from.IsVisible <- false
-            }
 
 open type Fabulous.Avalonia.View
 
 module TransitioningContentControlPage =
     type Model =
-        { Images: Bitmap list
-          SelectedImage: Bitmap
-          PageTransitions: PageTransition list
-          SelectedTransition: PageTransition
-          Duration: float option
+        { Images: string list
+          SelectedImage: string
+          PageTransition: IPageTransition
+          IsReverse: bool
+          Transitions: string list
+          Duration: float
           ClipToBounds: bool }
 
     type Msg =
@@ -99,60 +29,23 @@ module TransitioningContentControlPage =
         | TransitionChanged of SelectionChangedEventArgs
         | DurationChanged of float option
         | ClipToBoundsChanged of bool
-        | TransitionsUpdated of PageTransition list
+        | ReverseChanged of bool
 
-    type CmdMsg = SettingUpTransitions of PageTransition list * int
-
-    let updateTransitions (pageTransitions: PageTransition list) duration =
-        let transitions =
-            pageTransitions
-            |> List.mapi(fun index value ->
-                match index with
-                | 1 -> value.Transition <- CrossFade(TimeSpan.FromMilliseconds(duration))
-                | 2 -> value.Transition <- PageSlide(TimeSpan.FromMilliseconds(duration), PageSlide.SlideAxis.Horizontal) :> IPageTransition
-                | 3 -> value.Transition <- PageSlide(TimeSpan.FromMilliseconds(duration), PageSlide.SlideAxis.Vertical) :> IPageTransition
-                | 4 ->
-                    let compositeTransition = CompositePageTransition()
-                    compositeTransition.PageTransitions.Add(pageTransitions[1].Transition)
-                    compositeTransition.PageTransitions.Add(pageTransitions[2].Transition)
-                    compositeTransition.PageTransitions.Add(pageTransitions[3].Transition)
-                    value.Transition <- compositeTransition :> IPageTransition
-                | 5 -> value.Transition <- CustomTransition(TimeSpan.FromMilliseconds(duration)) :> IPageTransition
-                | _ -> value.Transition <- null
-
-                value)
-
-        TransitionsUpdated transitions
-
+    type CmdMsg = | NoCmdMsg
 
     let mapCmdMsgToCmd cmdMsg =
         match cmdMsg with
-        | SettingUpTransitions(pageTransitions, i) -> Cmd.ofMsg(updateTransitions pageTransitions i)
-
-    let pageTransitions =
-        [ PageTransition("None")
-          PageTransition("CrossFade")
-          PageTransition("Slide horizontally")
-          PageTransition("Slide vertically")
-          PageTransition("Composite")
-          PageTransition("Custom") ]
+        | NoCmdMsg -> Cmd.none
 
     let init () =
-        let images = [ "fabulous-icon.png"; "fsharp-icon.png"; "fabulous-icon.png" ]
-
-        let images =
-            images
-            |> List.map(fun image ->
-                let path = $"avares://Gallery/Assets/Icons/{image}"
-                ImageSource.fromString path)
-
-        { Images = images
-          SelectedImage = images[0]
-          SelectedTransition = pageTransitions[0]
-          PageTransitions = pageTransitions
-          Duration = Some 250.
+        { Images = [ "fabulous-icon.png"; "fsharp-icon.png"; "fabulous-icon.png" ]
+          SelectedImage = "fabulous-icon.png"
+          PageTransition = PageSlide(TimeSpan.FromSeconds(250.), PageSlide.SlideAxis.Horizontal)
+          Transitions = [ "Slide"; "CrossFade"; "3D Rotation"; "Composite" ]
+          Duration = 250.
+          IsReverse = false
           ClipToBounds = false },
-        [ SettingUpTransitions(pageTransitions, 500) ]
+        []
 
     let update msg model =
         match msg with
@@ -161,7 +54,7 @@ module TransitioningContentControlPage =
             let index = (index + 1) % model.Images.Length
 
             { model with
-                SelectedImage = model.Images[index] },
+                SelectedImage = model.Images.[index] },
             []
 
         | PrevImage ->
@@ -170,12 +63,12 @@ module TransitioningContentControlPage =
             let index = if index < 0 then model.Images.Length - 1 else index
 
             { model with
-                SelectedImage = model.Images[index] },
+                SelectedImage = model.Images.[index] },
             []
 
         | DurationChanged duration ->
             let duration = duration |> Option.defaultValue 500
-            { model with Duration = Some duration }, [ SettingUpTransitions(model.PageTransitions, int duration) ]
+            { model with Duration = duration }, []
 
         | ClipToBoundsChanged clipToBounds ->
             { model with
@@ -185,22 +78,30 @@ module TransitioningContentControlPage =
         | TransitionChanged selection ->
             let control = selection.Source :?> ComboBox
             let selectedItem = control.SelectedIndex
-            let selection = control.Items[selectedItem] :?> string
+            let selection = control.Items.[selectedItem] :?> string
 
-            let selectedTransition =
-                model.PageTransitions |> List.tryFind(fun x -> x.DisplayTitle = selection)
+            let transition =
+                match selection with
+                | "Slide" -> PageSlide(TimeSpan.FromMilliseconds(model.Duration), PageSlide.SlideAxis.Horizontal) :> IPageTransition
+                | "CrossFade" -> CrossFade(TimeSpan.FromSeconds(model.Duration))
+                | "3D Rotation" -> Rotate3DTransition(TimeSpan.FromSeconds(model.Duration), PageSlide.SlideAxis.Horizontal)
+                | "Composite" ->
+                    let crossFade = CrossFade(TimeSpan.FromSeconds(model.Duration))
+                    crossFade.FadeInEasing <- BounceEaseIn()
+                    crossFade.FadeOutEasing <- BounceEaseOut()
 
-            match selectedTransition with
-            | None -> model, []
-            | Some value ->
-                { model with
-                    SelectedTransition = selectedTransition.Value },
-                []
+                    let compositePageTransition = CompositePageTransition()
+                    compositePageTransition.PageTransitions.Add(Rotate3DTransition(TimeSpan.FromSeconds(model.Duration), PageSlide.SlideAxis.Horizontal))
+                    compositePageTransition.PageTransitions.Add(crossFade)
+                    compositePageTransition
 
-        | TransitionsUpdated pageTransitions ->
+                | _ -> PageSlide(TimeSpan.FromSeconds(model.Duration), PageSlide.SlideAxis.Horizontal)
+
             { model with
-                PageTransitions = pageTransitions },
+                PageTransition = transition },
             []
+
+        | ReverseChanged isReverse -> { model with IsReverse = isReverse }, []
 
     let program =
         Program.statefulWithCmdMsg init update mapCmdMsgToCmd
@@ -227,7 +128,7 @@ module TransitioningContentControlPage =
                     (HStack(5.) {
                         HeaderedContentControl(
                             "Select a transition",
-                            ComboBox(model.PageTransitions |> List.map(_.DisplayTitle), (fun x -> TextBlock(x)))
+                            ComboBox(model.Transitions, (fun x -> TextBlock(x)))
                                 .selectedIndex(0)
                                 .onSelectionChanged(TransitionChanged)
                                 .verticalAlignment(VerticalAlignment.Center)
@@ -235,7 +136,7 @@ module TransitioningContentControlPage =
 
                         HeaderedContentControl(
                             "Duration",
-                            NumericUpDown(100., 1000., model.Duration, DurationChanged)
+                            NumericUpDown(100., 1000., Some model.Duration, DurationChanged)
                                 .increment(250.)
                                 .verticalAlignment(VerticalAlignment.Center)
                         )
@@ -263,9 +164,14 @@ module TransitioningContentControlPage =
 
                 Button(">", NextImage).dock(Dock.Right)
 
+                CheckBox("Reverse", model.IsReverse, ReverseChanged)
+                    .dock(Dock.Bottom)
+                    .margin(5.)
+
                 Border(
-                    TransitioningContentControl(Image(model.SelectedImage))
-                        .pageTransition(model.SelectedTransition.Transition)
+                    TransitioningContentControl(Image($"avares://Gallery/Assets/Icons/{model.SelectedImage}"))
+                        .pageTransition(model.PageTransition)
+                        .isTransitionReversed(model.IsReverse)
                         .size(200., 200.)
                 )
                     .margin(5.)