+++ templates/content/blank/App.fs
@@ -26,7 +26,7 @@ module App =
             do! Async.Sleep 200
             return TimedTick
         }
-        |> Cmd.ofAsyncMsg
+        |> Cmd.OfAsync.msg
 
     let init () = initModel, Cmd.none