+++ templates/content/multi/NewApp/App.fs
@@ -27,7 +27,7 @@ module App =
             do! Async.Sleep 200
             return TimedTick
         }
-        |> Cmd.ofAsyncMsg
+        |> Cmd.OfAsync.msg
 
     let init () = initModel, Cmd.none